feat: prevent api calls when user is not logged in

+++ src/main/events/library/remove-game-from-library.ts
@@ -19,7 +19,7 @@ const removeGameFromLibrary = async (
 const removeRemoveGameFromLibrary = async (gameId: number) => {
   const game = await gameRepository.findOne({ where: { id: gameId } });
 
-  if (game?.remoteId) {
+  if (game?.remoteId && HydraApi.isLoggedIn()) {
     HydraApi.delete(`/games/${game.remoteId}`);
   }
 };
